third_party/WebKit/LayoutTests/external/wpt/pointerevents/compat/pointerevent_touch-action_two-finger_interaction-manual-expected.txt

This is a testharness.js-based test.
FAIL touch two-finger pan on 'touch-action: pan-x pan-y' assert_equals: expected "pointerdown@black, pointerdown@black, pointerup@black, pointerup@black" but got "pointerdown@black, pointerdown@black, pointercancel@black, pointercancel@black"
PASS touch two-finger pan on 'touch-action: pinch-zoom' 
PASS PointerEvent Automation 
Harness: the test ran to completion.
